+++ lib/modules/navigation/navigation_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: navigation
-// Updated at: 2026-05-06 19:17:57
-// Commit: test(navigation): implement widget integration tests for navigation
+// Updated at: 2026-05-09 21:27:50
+// Commit: feat(navigation): implement shimmer loading skeleton in navigation
 
 class navigationComponentState {
-  final String commitRef = '955';
-  final String lastMessage = 'test(navigation): implement widget integration tests for navigation';
-  final DateTime timestamp = DateTime.parse('2026-05-06T19:17:57+05:00');
+  final String commitRef = '985';
+  final String lastMessage = 'feat(navigation): implement shimmer loading skeleton in navigation';
+  final DateTime timestamp = DateTime.parse('2026-05-09T21:27:50+05:00');
 }